Replace TARGET_PRODUCT with TARGET_KERNEL_MEDIATEK_PRODUCT.
The MediaTek build system used the TARGET_PRODUCT variable to know the

specific product from the MediaTek tree that the kernel is built for and

thus set paths and other variables to the appropriate values. However,

the Android build system uses TARGET_PRODUCT for the general Android

product being built, which does not have to be the same as the one from

the MediaTek kernel tree. Therefore, TARGET_PRODUCT in the MediaTek

build system for the kernel was replaced with

TARGET_KERNEL_MEDIATEK_PRODUCT.

Although TARGET_PRODUCT is used in other areas of the MediaTek build

system only those affecting the build of the kernel were modified; the

rest are irrelevant, so they were just ignored.

Also, the ahong89_wet_jb2 is the only MediaTek custom product

supported by this kernel tree, so it was set as the

TARGET_KERNEL_MEDIATEK_PRODUCT default value.